+++ revocation/ocsp/ocsp.go
@@ -99,7 +99,7 @@ func CheckStatus(opts Options) ([]*result.CertRevocationResult, error) {
 		// Assume cert chain is accurate and next cert in chain is the issuer
 		go func(i int, cert *x509.Certificate) {
 			defer wg.Done()
-			certResults[i] = CertCheckStatus(cert, opts.CertChain[i+1], opts)
+			certResults[i] = certCheckStatus(cert, opts.CertChain[i+1], opts)
 		}(i, cert)
 	}
 	// Last is root cert, which will never be revoked by OCSP
@@ -115,16 +115,15 @@ func CheckStatus(opts Options) ([]*result.CertRevocationResult, error) {
 	return certResults, nil
 }
 
-// CertCheckStatus checks the revocation status of a certificate using OCSP
-func CertCheckStatus(cert, issuer *x509.Certificate, opts Options) *result.CertRevocationResult {
-	if !SupportOCSP(cert) {
+func certCheckStatus(cert, issuer *x509.Certificate, opts Options) *result.CertRevocationResult {
+	ocspURLs := cert.OCSPServer
+	if len(ocspURLs) == 0 {
 		// OCSP not enabled for this certificate.
 		return &result.CertRevocationResult{
 			Result:        result.ResultNonRevokable,
 			ServerResults: []*result.ServerResult{toServerResult("", NoServerError{})},
 		}
 	}
-	ocspURLs := cert.OCSPServer
 
 	serverResults := make([]*result.ServerResult, len(ocspURLs))
 	for serverIndex, server := range ocspURLs {
@@ -140,11 +139,6 @@ func CertCheckStatus(cert, issuer *x509.Certificate, opts Options) *result.CertR
 		serverResults[serverIndex] = serverResult
 	}
 	return serverResultsToCertRevocationResult(serverResults)
-}
-
-// SupportOCSP returns true if the certificate supports OCSP.
-func SupportOCSP(cert *x509.Certificate) bool {
-	return len(cert.OCSPServer) > 0
 }
 
 func checkStatusFromServer(cert, issuer *x509.Certificate, server string, opts Options) *result.ServerResult {
