+++ revocation/result/errors.go
@@ -31,26 +31,3 @@ func (e InvalidChainError) Error() string {
 	}
 	return msg
 }
-
-// OCSPFallbackErro is returned when the OCSP check result is of unknown status
-// and falls back to CRL
-type OCSPFallbackError struct {
-	// OCSPErr is the error that occurred during the OCSP check
-	OCSPErr error
-
-	// CRLErr is the error that occurred during the CRL check
-	CRLErr error
-}
-
-func (e OCSPFallbackError) Error() string {
-	msg := "the OCSP check result is of unknown status; fallback to CRL"
-	if e.OCSPErr != nil {
-		msg += fmt.Sprintf("; OCSP error: %v", e.OCSPErr)
-	}
-
-	if e.CRLErr != nil {
-		msg += fmt.Sprintf("; CRL error: %v", e.CRLErr)
-	}
-
-	return msg
-}
